backend/requirements.txt
setuptools
wheel
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
python-dateutil==2.8.2
joblib==1.3.1
mysql-connector-python==9.0.0
gunicorn==21.2.0
google-generativeai==0.3.0
